+++ Graphs/Tools/grasp_laplacian_standard.m
@@ -1,15 +1,21 @@
-%Constructs the standard Laplacian of a non directed graph.
+%Constructs the standard Laplacian of a non directed graph. If the graph is
+%directed, with non-negative weights, use Fan Chung combinatorial Laplacian
+%[Laplacians and the Cheeger Inequality for Directed Graphs].
 %
 %   L = GRASP_LAPLACIAN_STANDARD(graph) returns the Laplacian of the
-%   graph provided using its adjacency matrix.
+%       graph provided using its adjacency matrix.
 %
 % Authors:
 %  - Benjamin Girault <benjamin.girault@ens-lyon.fr>
+%  - Benjamin Girault <benjamin.girault@usc.edu>
 
 % Copyright Benjamin Girault, École Normale Supérieure de Lyon, FRANCE /
 % Inria, FRANCE (2015-11-01)
+% Copyright Benjamin Girault, University of Sourthern California, Los
+% Angeles, California, USA (2018-2019)
 % 
 % benjamin.girault@ens-lyon.fr
+% benjamin.girault@usc.edu
 % 
 % This software is a computer program whose purpose is to provide a Matlab
 % / Octave toolbox for handling and displaying graph signals.
@@ -41,11 +47,26 @@
 % knowledge of the CeCILL license and that you accept its terms.
 
 function L = grasp_laplacian_standard(graph)
-    %% Checks
-    if grasp_is_directed(graph)
-        error('Error: directed graph!');
+    %% Check for self-loops
+    if sum(abs(diag(graph.A))) > 0
+        error('GraSP:LaplacianStandard:UnsupportedSelfLoops', 'Error: unsupported graph with self-loops!');
     end
     
-    %% Laplacian
-    L = grasp_degrees(graph) - grasp_adjacency_matrix(graph);
+    %% Compute L
+    if grasp_is_directed(graph)
+        %% Check that we can convert the adjacency matrix into a random walk matrix P
+        if sum(graph.A(:) < 0) > 0
+            error('GraSP:LaplacianStandard:UnsupportedNegativeWeights', 'Error: directed graph with negative weights!');
+        end
+        
+        %% Directed graph Laplacian deriving from F. Chung definition
+        Dout = grasp_degrees(graph);
+        P = Dout ^ -1 * graph.A;
+        [pi, ~] = eigs(P', 1, 1); % Closest left eigenvector of P to eigenvalue 1 (stationary distribution of the RW)
+        Pi = diag(pi);
+        L = Pi - (Pi * P + P' * Pi) / 2;
+    else
+        %% Classic Combinatorial Laplacian for undirected graph
+        L = grasp_degrees(graph) - grasp_adjacency_matrix(graph);
+    end
 end